requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
python-dotenv==0.14.0
six==1.15.0
Werkzeug==1.0.1
